fix accrual type to float

+++ internal/models/order/order.go
@@ -9,7 +9,7 @@ type Order struct {
 	UserLogin  string     `json:"-"`
 	Number     Number     `json:"number"`
 	Status     Status     `json:"status"`
-	Accrual    *int64     `json:"accrual,omitempty"`
+	Accrual    *float64   `json:"accrual,omitempty"`
 	UploadedAt *time.Time `json:"uploaded_at"`
 }
